requirements.txt
google-generativeai
pyautogui
pygetwindow
Pillow
pywin32; sys_platform == "win32"
uiautomation; sys_platform == "win32"

# RAG dependencies
chromadb
sentence-transformers
PyMuPDF
pyparsing

# OCR dependencies for text highlighting
pytesseract

# OmniParser dependencies
torch
easyocr
torchvision
supervision==0.18.0
openai==1.3.5
transformers==4.51.1
ultralytics==8.3.70
azure-identity
numpy==1.26.4
opencv-python
opencv-python-headless
gradio
dill
accelerate
timm
einops==0.8.0
paddlepaddle
paddleocr
ruff==0.6.7
pre-commit==3.8.0
pytest==8.3.3
pytest-asyncio==0.23.6
pyautogui==0.9.54
streamlit>=1.38.0
anthropic[bedrock,vertex]>=0.37.1
jsonschema==4.22.0
boto3>=1.28.57
google-auth<3,>=2
screeninfo
uiautomation
dashscope
groq
